Treat files as new flag for first seen timestamp
This flag will cause the watcher to look at the ctime of a file if it is
true. This will not work on Windows systems if the file is simply moved
from one queue to the other. It will work on Linux systems.

Default is false: the timestamp of first seen will be the `now` of when
the watch first sees the file in the walk.